Create methods to retrieve ride by ID from ride list
requestedID is the id as according to ride.ts's OID that is retrieved
from a button click from a button on a ride. requestRideID() takes that
field. retrieveRide() takes the OID and calls the listService to
initiate a GET request for that ride by ID.

As of this commit, that service is hard coded to not use the
retrieveRide OID but instead a particular explicit OID. Once we are
more certain about how GET requests work, the service will be
paramaterized.

#4